Move pattern related codes to pattern.py from operation.py; Add a new API called apply() to Schedule class
